requirement.txt
Flask==2.2.2
Flask-Cors==3.0.10
pyzmq==26.4.0
debugpy==1.8.14
Werkzeug==2.3.7